requirements.txt
alpaca-py==0.40.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
Credit-Spreader==0.1
cryptography==44.0.3
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
mongoengine==0.29.1
msgpack==1.1.0
numpy==2.2.5
openai==1.78.1
pandas==2.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pymongo==4.13.0
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sseclient-py==1.8.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websockets==15.0.1
